feat: copy common 404.md source file to projects
