+++ src/rocm_docs/data/util_pages/404.md
@@ -1,0 +1,6 @@
+# 404 Page Not Found
+
+Page could not be found.
+
+Please use the links from the sidebar to find what
+you are looking for.
